Track per-file bytes completed data. For now, this is all 0's.
This change allows us to round trip per-file bytes completed data
through the lifetime of a torrent, including persistence in the boltdb
so that it is persisted across restarts.

Extend the Spec we use for boltdb. Teach the boltdb Spec we use how to
handle a PerFileBytesDownloaded []int64 that represents how many bytes
of each file have been downloaded. We teach session loading how to
work with the new field as well, although it currently isn't used
beyond that. This will give us a foothold for having the torrent
struct manage an additional set of stats.

Teach resumer.Stats how to handle per-file downloaded bytes. Add a new
field PerFileBytesDownloaded to resumer.Stats and pass that around
when loading a torrent from existing state. This still isn't used by
the torrent, but will now be available to it.

Add perFileBytesDownloaded to the torrent struct. This field is a map
of pathname to bytes downloaded for that pathname/file. When restoring
a torrent from a previous session, we use the resumer stats to
populate the data. The resumer data is a slice of int64's, so we make
the assumption that the ordering is the same as that of the metadata
Files slice and index through it. Because this will be accessed
concurrently, we need to add a mutex for it. Wrap existing accesses to
it with that new mutex.

We also ensure that when a new torrent is added, we initialize the
metrics counter for each file. This ensures that we shouldn't need to
guard all other read operations with verifying that the map entries
exist.

Ensure that per-file downloaded bytes are updated in the db. Teach
torrent.updateStats how to update the per-file downloaded bytes
metadata that we're wiring up to be tracked.