Add a Files method to the Torrent object.
This method returns a slice of File objects, newly defined for this
purpose. Each File has a Path and Stats method which allow querying
for the file path and per-file size and download completion in
bytes. This method is only usable when metainfo is available (torrent
in one of the running states) as it relies on piece data to build up
the notion of completeness. The piece done state, which should mirror
the bitfield completion status is how we determine whether or not to
count bytes as completed for the file sections in that particular
piece. Padding files are ignored.